python_server/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python_server"
version = "0.1.0"
description = "FastAPI service for hello-all-worlds"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.0",
  "pydantic-settings>=2.0",
  "python-dotenv>=1.0",
  "python_db @ file:../python_db",
  "python_utils @ file:../python_utils",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*"]